Fix bugs relating to reactor retirement:
Previously, reactors would fail to record their last entry in TimeSeriesPower
table due to retirement guards that prevented the code from executing.  This
fixes that.  There may still be a corner case that isn't handled quite right
if a reactor's refueling cycle ends exactly on the time step that it retires,
but I'm not sure.

When a reactor approaches retirement, it reduces its fuel ordering to only
acquire enough to last until retirement.  Previously, this didn't account for
the reactor potentially being short on fuel (i.e. the core wasn't full) and so
could under-order fresh fuel in certain circumstances.  This has been fixed.