Prevent start of a new cycle if core has not been discharged
The if-condition leading to `cycle_step` being reset to 0 and subsequently
initiating the start of a new cycle has been changed to avoid erroneous
behaviour.

If the spent fuel inventory is of size zero or if it is full in between
time `cycle_time` and `cycle_time + refuel_time`, then the core cannot
be discharged in `Reactor::Tick()` and `discharged` remains set to false.

In `Reactor::Tock()`, this leads to `core.count() == n_assem_core`
evaluating to true and if `cycle_step` is larger or equal than
`cycle_time + refuel_time`, then `cycle_step` gets reset to 0 and a new
cycle will (wrongly) start with a full core composed at least partially
of spent fuel.

The additional boolean expression prevents this behaviour by checking if
the core has been discharged before resetting `cycle_step`.